refactor: rename into beine_dienen_legs and update dependency
