+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.5)
 
-project(bein_bridge)
+project(beine_dienen_legs)
 
 if(NOT CMAKE_C_STANDARD)
   set(CMAKE_C_STANDARD 99)
@@ -15,7 +15,7 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 endif()
 
 find_package(ament_cmake REQUIRED)
-find_package(bein_interfaces REQUIRED)
+find_package(beine_interfaces REQUIRED)
 find_package(musen REQUIRED)
 find_package(rclcpp REQUIRED)
 
@@ -28,7 +28,7 @@ target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
 
-ament_target_dependencies(${PROJECT_NAME} bein_interfaces musen rclcpp)
+ament_target_dependencies(${PROJECT_NAME} beine_interfaces musen rclcpp)
 
 install(DIRECTORY "include" DESTINATION ".")
 
@@ -54,7 +54,8 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 endif()
 
-ament_export_dependencies(bein_interfaces musen rclcpp)
+ament_export_dependencies(beine_interfaces musen rclcpp)
 ament_export_include_directories("include")
 ament_export_libraries(${PROJECT_NAME})
+
 ament_package()
