+++ README.md
@@ -1,11 +1,11 @@
-# Bein Bridge
+# Beine Dienen Legs
 
-[![latest version](https://img.shields.io/github/v/release/threeal/bein_bridge)](https://github.com/threeal/bein_bridge/releases/)
-[![milestone](https://img.shields.io/github/milestones/progress/threeal/bein_bridge/1?label=milestone)](https://github.com/threeal/bein_bridge/milestone/1)
-[![license](https://img.shields.io/github/license/threeal/bein_bridge)](./LICENSE)
-[![test status](https://img.shields.io/github/workflow/status/threeal/bein_bridge/Build%20and%20Test?label=test)](https://github.com/threeal/bein_bridge/actions)
+[![latest version](https://img.shields.io/github/v/release/threeal/beine_dienen_legs)](https://github.com/threeal/beine_dienen_legs/releases/)
+[![milestone](https://img.shields.io/github/milestones/progress/threeal/beine_dienen_legs/1?label=milestone)](https://github.com/threeal/beine_dienen_legs/milestone/1)
+[![license](https://img.shields.io/github/license/threeal/beine_dienen_legs)](./LICENSE)
+[![test status](https://img.shields.io/github/workflow/status/threeal/beine_dienen_legs/Build%20and%20Test?label=test)](https://github.com/threeal/beine_dienen_legs/actions)
 
-This package provides bridge programs that enable communication between [ROS 2](https://docs.ros.org/en/foxy/index.html) and Bein legs reinforcement.
+This package provides a [Dienen](https://github.com/threeal/proposal-ta-simulasi-robot) assistive robot's leg support implementation of [Beine](https://github.com/threeal/beine) leg support project.
 
 ## License
