+++ src/listeners/voice_listener.cpp
@@ -18,12 +18,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <bein_bridge/listeners/voice_listener.hpp>
+#include <beine_dienen_legs/listeners/voice_listener.hpp>
 
 #include <memory>
 #include <string>
 
-namespace bein_bridge
+namespace beine_dienen_legs
 {
 
 using namespace std::chrono_literals;
@@ -59,4 +59,4 @@ void VoiceListener::listen_process()
   }
 }
 
-}  // namespace bein_bridge
+}  // namespace beine_dienen_legs
